+++ internal/commands/database/session/session.go
@@ -1,0 +1,16 @@
+package databasesession
+
+import (
+	"github.com/UpCloudLtd/upcloud-cli/v2/internal/commands"
+)
+
+// BaseSessionCommand creates the base "session" command
+func BaseSessionCommand() commands.Command {
+	return &databaseSessionCommand{
+		commands.New("session", "Manage database sessions"),
+	}
+}
+
+type databaseSessionCommand struct {
+	*commands.BaseCommand
+}